__download: changed to determinant tempfile
Instead of having abandoned tempfiles from previously failed cdist runs,
this type will use a previctive filename. If failed, the next cdist run
takes the same file instead of creating a new one.

This works as via the cdist principles, there can't be multiple
objects/downloads for a single file (as it's the object id).