Use "assert pattern" instead of if ! ...
Replace occurrences of "if ! ..." with "... | { handler }" in cases when it is
not really used as a conditional but to check for a requirement/prerequisite and
handling an error condition in the block.

This way it is clearer that the code is not a conditional but instead it is to
handle an unexpected condition.

Such blocks usually end with "exit 1" :-)