[type/__apt_sources] Add support for old versions without sources.list.d
Support for sources.list.d was introduced in APT 0.6.43.
Debian < 4.0 and Ubuntu < 6.06 thus do not support it.

Instead we can use a __block in /etc/apt/sources.list.

While at it, this commit also gets rid of the non-necessary temporary file.

Distribution codename detection does not work for these old versions
which neither have LSB nor os-release(5), so the --distribution has
to be provided manually.

Also, in the very unlikely case that someone upgrades from such an
old system to a newer version, the __apt_source in /etc/apt/sources.list
would remain.
I think we can live with this. Patches welcome :D