[type/__group] Remove not_supported error from gshadow explorer
Having gshadow support is not a requirement for using the __group type, so it
should not produce a hard error.
Further, the implementation was bugged, because the error would only appear on
systems which don't support gshadow _and_ don't have getent(1) which makes no
sense.

stderr is silenced for getent(1) because the call will error on operating
systems without gshadow support (e.g. BSDs), but this is fine.