[explorer/os] Adopt rc_getvar() to parse os-release(5)
Adopt the rc_getvar() function to parse the os-release(5) file instead of using
eval to parse a file which could in theory contain (malicious) code.

Further, get rid of an ugly regex to match SuSE derivatives which fails to
evaluate in some implementations of grep(1) due to the
alternatives (e.g. OpenBSD).