[explorer/cpu_cores] Update detection for NetBSD, fallback for old OpenBSD
NetBSD also has hw.ncpuonline like OpenBSD (in fact, OpenBSD imported
the code for it from NetBSD).

Let's merge the code for NetBSD and OpenBSD instead of FreeBSD.

On OpenBSD sysctl(8) never returns a non-zero exit status, so `grep .' is
used to detect success and fall back to hw.ncpu.

https://marc.info/?l=openbsd-cvs&m=153135863131342&w=2