[type/__sed] Fix for NetBSD (-f -)
NetBSD sed(1) does not accept -f - to read the script from stdin:

    sed: -: No such file or directory

Since the script has to be copied to code-remote anyway (for the `--script -`
case), instead of putting it into a here-document we can pass it into sed -e
instead.
I decided to go with multiple -e options (one per line), but passing the whole
script into a single -e should work as well.