[type/__start_on_boot] Speed up state explorer on OpenBSD
Instead of running `rcctl ls on` and grep(1)ing for the service of interest, we
can just query the service's status.  `rcctl ls on` does the same thing, but it
queries the status for all services which is unnecessary for this case.

Further, the grep(1) (which interpreted the service name as a regular expression)
is not necessary anymore.  Another issue with the grep was that it printed the
service name if it matched, breaking `$state_is` in `manifest` and
`gencode-remote`.